+ version 0.1.06 + replaced float with double to support ARM (DUE) too

+++ libraries/FastMap/FastMap.cpp
@@ -1,12 +1,13 @@
 //
 //    FILE: FastMap.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.05
+// VERSION: 0.1.06
 // PURPOSE: class implementation of map function - library for Arduino
 //     URL: http://forum.arduino.cc/index.php?topic=276194
 //
 // HISTORY:
-// 0.1.05 stripped of bit mask experimental code
+// 0.1.06 2015-03-08 replaced double by double (support ARM)
+// 0.1.05 2014-11-02 stripped of bit mask experimental code
 // 0.1.04 add back() - the inverse map
 //        tested with bit mask for constrain code (Perfomance was killed)
 // 0.1.03 proper name
@@ -21,7 +22,7 @@
 //
 // PUBLIC
 //
-void FastMap::init(float in_min, float in_max, float out_min, float out_max)
+void FastMap::init(double in_min, double in_max, double out_min, double out_max)
 {
     _in_min = in_min;
     _in_max = in_max;
@@ -35,20 +36,20 @@ void FastMap::init(float in_min, float in_max, float out_min, float out_max)
     _backbase = in_min - out_min * _backfactor;
 }
 
-float FastMap::constrainedMap(float value)
+double FastMap::constrainedMap(double value)
 {
     if (value <= _in_min) return _out_min;
     if (value >= _in_max) return _out_max;
     return this->map(value);
 }
 
-float FastMap::lowerConstrainedMap(float value)
+double FastMap::lowerConstrainedMap(double value)
 {
     if (value <= _in_min) return _out_min;
     return this->map(value);
 }
 
-float FastMap::upperConstrainedMap(float value)
+double FastMap::upperConstrainedMap(double value)
 {
     if (value >= _in_max) return _out_max;
     return this->map(value);
